+++ product/pom.xml
@@ -17,7 +17,7 @@
         <java.version>17</java.version>
         <sonar.organization>xandervedder</sonar.organization>
         <sonar.host.url>https://sonarcloud.io/</sonar.host.url>
-        <sonar.exclusions>**/ProductApplication.java</sonar.exclusions>
+        <sonar.exclusions>**/ProductApplication.java, **/infrastructure.**, **/core/common.**</sonar.exclusions>
     </properties>
     <dependencies>
         <dependency>
@@ -77,6 +77,8 @@
                         <configuration>
                             <excludes>
                                 <exclude>**/ProductApplication.**</exclude>
+                                <exclude>**/infrastructure.**</exclude>
+                                <exclude>**/core/common.**</exclude>
                             </excludes>
                             <rules>
                                 <rule>
