+++ product/src/test/java/nl/softwarestrijders/waiter/product/domain/NutritionalValueTest.java
@@ -1,5 +1,6 @@
 package nl.softwarestrijders.waiter.product.domain;
 
+import nl.softwarestrijders.waiter.product.core.domain.NutritionalValue;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
